Although 'Unanimous Consent' appears to be a well defined and used in the US Senate, Canadian Parliament and Australian Senate, it was causing some confusion.  After some consultation with Mark Murray, change this to 'without objection' since often times a plain-speaking term is preferable to a regionally used term.
Also, clarify that this procedure is to be used when for more mundane
matters that need a sanity check, but don't need the whole, ponderous
voting proceedure that more difficult issues require.  Core members
that read email in any given 48 hour period are trusted enough to know
the difference and to provide the sanity check as necessary.

Reviewed by: markm