Move from using the socket reference count to the file reference count to prevent sockets from being garbage collected during socket-specific system calls.  This is the same approach used in most VFS-specific system calls, as well as generic file descriptor system calls such as read() and write().
To do this, add a utility function getsock(), which is logically
identical to getvnode() used for the same purpose in VFS.  Unlike
fgetsock(), it returns with the file reference count elevated, but
no bump of the socket reference count.  Replace matching calls to
fputsock() with fdrop().

This change is made to all socket system calls other than
sendfile() and accept(), but the approach should be applicable to
those system calls also.

This shaves about four mutex operations off of each of these
system calls, including send() and recv() variants, adding about
1% to pps on minimal UDP packets for UP using netblast, and 4% on
SMP.

Reviewed by:	pjd


git-svn-id: svn+ssh://svn.freebsd.org/base/head@136930 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f