Although being recorded as MFC of the r177369, this change was introduced for the KSE case in the r177083. Both were done by jeff, MFC is needed due to recent cpuset(2) commit to RELENG_7.
r177083:
KSE may free a thread that was never actually forked. This will leave
td_cpuset NULL. Check for this condition before dereferencing the
cpuset.

r177369:
Restore the NULL check for td_cpuset. This can happen if a partially
constructed thread was torn down as is the case when we fail to allocate
a kernel stack.