Make ureadc() warn when holding any locks, just like uiomove().
A couple of months ago I was quite impressed, because when I was writing
code, I discovered that uiomove() would not allow any locks to be held,
while ureadc() did, mainly because ureadc() is implemented using the
same building blocks as uiomove().

Let's see if this triggers any aditional witness warnings on our source
tree.

Reviewed by:	atillio


git-svn-id: svn+ssh://svn.freebsd.org/base/head@182399 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f