Introduce libefi -- a library around EFI runtime services and protocols. This first commit brings 3 functions for enumerating, retrieving, adding, removing and modifying EFI variables. The immediate use of these include the insertion of a new boot option as part of the installation process.
This library uses ioctl(2) requests implemented by io(4) to pass the
requests down through the kernel to EFI. These ioctl requests are only
implemented on ia64, so libefi is currently only enabled on ia64. The
interface is generic and io(4) on mad64/i386 can easily be taught to
handle these once EFI support has been added to the kernel there.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@203181 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f