Fix an error in the EXAMPLES section of getopt(1), which is based on the same fix present in NetBSD.
Note: the getopt man page contains more antique information like this.
An overhaul of the man page and/or sync with NetBSD would be the right
thing to do. But since this is out of the scope of the PR, I'll leave
it as it is for now.

PR:             docs/133118
Submitted by:   Oleg A. Mamontov (oleg at mamontov dot net)
Discussed with: jilles@
MFC after:      2 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@209772 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f