Merge r216333 and r216555 from the native pmap   When r207410 eliminated the acquisition and release of the page queues   lock from pmap_extract_and_hold(), it didn't take into account that   pmap_pte_quick() sometimes requires the page queues lock to be held.   This change reimplements pmap_extract_and_hold() such that it no   longer uses pmap_pte_quick(), and thus never requires the page queues   lock.
Merge r177525 from the native pmap
  Prevent the overflow in the calculation of the next page directory.
  The overflow causes the wraparound with consequent corruption of the
  (almost) whole address space mapping.

Strictly speaking, r177525 is not required by the Xen pmap because the
hypervisor steals the uppermost region of the normal kernel address
space.  I am nonetheless merging it in order to reduce the number of
unnecessary differences between the native and Xen pmap implementations.

Tested by:	sbruno


git-svn-id: svn+ssh://svn.freebsd.org/base/head@229007 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f