fbt_typoff_init: fix an off by one in determining required memory size
This issue would be silent most of the time, but if the requested memory
is a multiple of a page size, then accessing one element beyond the end
would lead to a kernel page fault.
Otherwise, the unlucky last type would just be inaccessible.

Reported by:	glebius
Tested by:	glebius
MFC after:	6 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@248640 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f