Fix changes made in r249408.
In some cases, kern_envp is set by the architecture code and env_pos does
not contain the length of the static kernel environment. In these cases
r249408 causes the kernel to discard the environment.

Fix this by updating the check for empty static env to *kern_envp != '\0'

Reported by:	np@


git-svn-id: svn+ssh://svn.freebsd.org/base/head@249436 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f