Change maxio to reflect variable hardware configurations.
If max_sg_length is 0, then we default to 16

If max_sg_length is less than CISS_MAX_SG_ELEMENTS, then
we will set round the value of max_sg_length to the nearest
power of 2 and use it to align maxio.

Else, we will use CISS_MAX_SG_ELEMENTS for our calculations.

Thanks to scottl for working me through the history and providing
the basis for this patch.

Submitted by:	scott
Obtained from:	Yahoo! Inc.
MFC after:	2 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@249977 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f