Import libucl into head
UCL is heavily infused by nginx configuration as the example of a convenient
configuration system. However, UCL is fully compatible with JSON format and is
able to parse json files.

UCL is used by pkg(8) for its configuration file as well for the manifest format
in packages, it will be used in base for the pkg boostrap (signature checking
and configuration file parsing.)

libucl has been developped and is maintained by vsevolod@


git-svn-id: svn+ssh://svn.freebsd.org/base/head@262398 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f